LayoutTests/accessibility/render-counter-text-expected.txt

This tests that the text for RenderCounter is exposed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS line1.childAtIndex(0).stringValue is 'AXValue: 1'
PASS line1.childAtIndex(1).stringValue is 'AXValue: :'
PASS line1.childAtIndex(2).stringValue is 'AXValue: line1'
PASS line2.childAtIndex(0).stringValue is 'AXValue: 2'
PASS line2.childAtIndex(1).stringValue is 'AXValue: :'
PASS line2.childAtIndex(2).stringValue is 'AXValue: line2'
PASS successfullyParsed is true

TEST COMPLETE
